+++ public/styles/skeleton.css
@@ -19,7 +19,7 @@
 .container section .skeleton-list li .dummy-img {
   width: 200px;
   height: 300px;
-  background: rgba(0, 0, 0, 0.5);
+  background: #f0f0f0;
   border-radius: 8px;
   margin-bottom: 4px;
 }
