+++ src/renders/movieList.ts
@@ -1,0 +1,67 @@
+import { Movie, Movies } from "../services/dto";
+import { removeMoreButton, renderMoreButton } from "./moreButton";
+
+const createMovieNode = (movie: Movie): DocumentFragment | null => {
+  const movieTemplate =
+    document.querySelector<HTMLTemplateElement>(`#movie-template`);
+  if (!movieTemplate) return null;
+
+  const movieFragment = movieTemplate.content.cloneNode(
+    true,
+  ) as DocumentFragment;
+
+  const movieItem = movieFragment.querySelector("li");
+  if (!movieItem) return null;
+  movieItem.dataset.movieId = String(movie.id);
+
+  const thumbnail = movieFragment.querySelector<HTMLImageElement>(".thumbnail");
+  if (!thumbnail) return null;
+  thumbnail.src =
+    `https://media.themoviedb.org/t/p/w220_and_h330_face` + movie.poster_path;
+  thumbnail.alt = movie.title;
+
+  const itemDesc = movieFragment.querySelector(".item-desc");
+
+  const rate = itemDesc?.querySelector("span");
+  if (!rate) return null;
+  rate.textContent = movie.vote_average.toString();
+
+  const title = itemDesc?.querySelector("strong");
+  if (!title) return null;
+  title.textContent = movie.title;
+
+  return movieFragment;
+};
+
+export const renderMovieList = (movies: Movies): void => {
+  const movieList = document.querySelector("#movie-list");
+
+  movies.results.forEach((movie: Movie) => {
+    const movieNode = createMovieNode(movie);
+    if (movieNode) {
+      movieList?.appendChild(movieNode);
+    }
+  });
+
+  if (movies.page === movies.total_pages) {
+    removeMoreButton();
+  } else {
+    renderMoreButton();
+  }
+};
+
+export const renderNoResult = () => {
+  const thumbnailList = document.querySelector(".thumbnail-list");
+  if (!thumbnailList) return;
+  const empty = '<p id="no-result">검색 결과가 없습니다.</p>';
+  thumbnailList.innerHTML = empty;
+
+  removeMoreButton();
+};
+
+export const removeMovieList = () => {
+  const movieList = document.querySelector<HTMLDivElement>("#movie-list");
+  if (!movieList) return;
+
+  movieList.replaceChildren();
+};
