+++ src/renders/skeleton.ts
@@ -1,0 +1,29 @@
+export const renderSkeleton = () => {
+  const skeleton = document.querySelector<HTMLDivElement>("#skeleton");
+  if (!skeleton) return;
+
+  const skeletonTemplate =
+    document.querySelector<HTMLTemplateElement>("#movie-template");
+  if (!skeletonTemplate) return null;
+
+  for (let i = 0; i < 20; i++) {
+    const skeletonCloneNode = skeletonTemplate.content.cloneNode(
+      true,
+    ) as DocumentFragment;
+    if (!skeletonCloneNode) return null;
+
+    skeleton.appendChild(skeletonCloneNode);
+  }
+};
+
+export const removeSkeleton = () => {
+  const skeleton = document.querySelector<HTMLDivElement>("#skeleton");
+  if (!skeleton) return;
+
+  skeleton.classList.add("animation");
+
+  setTimeout(() => {
+    skeleton.classList.remove("animation");
+    skeleton.replaceChildren();
+  }, 3000);
+};
